test_requirements.txt
apache-airflow==2.9.1
rabbit-bq-job-optimizer
apache-airflow-providers-google
google-cloud-bigquery
